Create new folder for ui-related stuff - move logos to new folder ui/logo
